Revert "feat(location): implement network scanning with iw cli (#426)"
This reverts commit 518891c724867b6645e382f7620b528c1830a50e.
