+++ location/daemon/Cargo.toml
@@ -20,7 +20,7 @@ orb-cellcom.workspace = true
 orb-endpoints.workspace = true
 orb-google-geolocation-api.workspace = true
 orb-info = { workspace = true, features = ["orb-id"] }
-orb-location-wifi.workspace = true
+orb-location-wpa-supplicant.workspace = true
 orb-security-utils = { workspace = true, features = [ "reqwest", "blocking" ] }
 reqwest = { workspace = true, features = ["blocking", "json"] }
 serde.workspace = true
