--- location/wifi/README.md
@@ -1,7 +1,0 @@
-# orb-location-wifi
-
-Implements all wifi related io.
-TODO: perhaps unify this code into the existing orb-wpa-supplicant crate.
-
-Important docs:
-- https://w1.fi/wpa_supplicant/devel/ctrl_iface_page.html
