+++ location/wpa-supp/README.md
@@ -1,0 +1,7 @@
+# orb-location-wpa-supplicant
+
+*Another* wpa-supplicant interface. TODO: unify this code into the existing
+orb-wpa-supplicant crate.
+
+Important docs:
+- https://w1.fi/wpa_supplicant/devel/ctrl_iface_page.html
